+++ src/pyhabitat/reporting.py
@@ -1,8 +1,10 @@
+# src/pyhabitat/report.py 
+from __future__ import annotations  
 import sys
 import logging
 from pathlib import Path
 
-from .environment import *
+from pyhabitat import environment as env
 def report(path=None, debug=False):
     """Print a comprehensive environment report.
 
@@ -19,47 +21,47 @@ def report(path=None, debug=False):
     print("\nCurrent Build Checks ")
     print("# // Based on hasattr(sys,..) and getattr(sys,..)")
     print("------------------------------")
-    print(f"in_repl(): {in_repl()}")
-    print(f"as_frozen(): {as_frozen()}")
-    print(f"as_pyinstaller(): {as_pyinstaller()}")
+    print(f"in_repl(): {env.in_repl()}")
+    print(f"as_frozen(): {env.as_frozen()}")
+    print(f"as_pyinstaller(): {env.as_pyinstaller()}")
     print("\nOperating System Checks")
     print("# // Based on platform.system()")
     print("------------------------------")
-    print(f"on_windows(): {on_windows()}")
-    print(f"on_apple(): {on_apple()}")
-    print(f"on_linux(): {on_linux()}")
-    print(f"on_wsl(): {on_wsl()}")
-    print(f"on_android(): {on_android()}")
-    print(f"on_termux(): {on_termux()}")
-    print(f"on_pydroid(): {on_pydroid()}")
-    print(f"on_ish_alpine(): {on_ish_alpine()}")
-    print(f"on_freebsd(): {on_freebsd()}")
+    print(f"on_windows(): {env.on_windows()}")
+    print(f"on_apple(): {env.on_apple()}")
+    print(f"on_linux(): {env.on_linux()}")
+    print(f"on_wsl(): {env.on_wsl()}")
+    print(f"on_android(): {env.on_android()}")
+    print(f"on_termux(): {env.on_termux()}")
+    print(f"on_pydroid(): {env.on_pydroid()}")
+    print(f"on_ish_alpine(): {env.on_ish_alpine()}")
+    print(f"on_freebsd(): {env.on_freebsd()}")
     print("\nCapability Checks")
     print("-------------------------")
-    print(f"tkinter_is_available(): {tkinter_is_available()}")
-    print(f"matplotlib_is_available_for_gui_plotting(): {matplotlib_is_available_for_gui_plotting()}")
-    print(f"matplotlib_is_available_for_headless_image_export(): {matplotlib_is_available_for_headless_image_export()}")
-    print(f"web_browser_is_available(): {web_browser_is_available()}")
-    print(f"interactive_terminal_is_available(): {interactive_terminal_is_available()}")
+    print(f"tkinter_is_available(): {env.tkinter_is_available()}")
+    print(f"matplotlib_is_available_for_gui_plotting(): {env.matplotlib_is_available_for_gui_plotting()}")
+    print(f"matplotlib_is_available_for_headless_image_export(): {env.matplotlib_is_available_for_headless_image_export()}")
+    print(f"web_browser_is_available(): {env.web_browser_is_available()}")
+    print(f"interactive_terminal_is_available(): {env.interactive_terminal_is_available()}")
     print("\nInterpreter Checks")
     print("# // Based on sys.executable()")
     print("-----------------------------")
-    print(f"interp_path(): {interp_path()}")
+    print(f"interp_path(): {env.interp_path()}")
     if debug:
         # Do these debug prints once to avoid redundant prints
         # Supress redundant prints explicity using suppress_debug=True, 
         # so that only unique information gets printed for each check, 
         # even when more than one use the same functions which include debugging logs.
-        #print(f"check_executable_path(interp_path(), debug=True)")
-        check_executable_path(interp_path(), debug=debug)    
-        #print(f"read_magic_bites(interp_path(), debug=True)")
-        read_magic_bytes(interp_path(), debug=debug)
-    print(f"is_elf(interp_path()): {is_elf(interp_path(), debug=debug, suppress_debug=True)}")
-    print(f"is_windows_portable_executable(interp_path()): {is_windows_portable_executable(interp_path(), debug=debug, suppress_debug=True)}")
-    print(f"is_macos_executable(interp_path()): {is_macos_executable(interp_path(), debug=debug, suppress_debug=True)}")
-    print(f"is_pyz(interp_path()): {is_pyz(interp_path(), debug=debug, suppress_debug=True)}")
-    print(f"is_pipx(interp_path()): {is_pipx(interp_path(), debug=debug, suppress_debug=True)}")
-    print(f"is_python_script(interp_path()): {is_python_script(interp_path(), debug=debug, suppress_debug=True)}")
+        #print(f"env.check_executable_path(env.interp_path(), debug=True)")
+        env.check_executable_path(env.interp_path(), debug=debug)    
+        #print(f"read_magic_bites(env.interp_path(), debug=True)")
+        env.read_magic_bytes(env.interp_path(), debug=debug)
+    print(f"is_elf(env.interp_path()): {env.is_elf(env.interp_path(), debug=debug, suppress_debug=True)}")
+    print(f"is_windows_portable_executable(env.interp_path()): {env.is_windows_portable_executable(env.interp_path(), debug=debug, suppress_debug=True)}")
+    print(f"is_macos_executable(env.interp_path()): {env.is_macos_executable(env.interp_path(), debug=debug, suppress_debug=True)}")
+    print(f"is_pyz(env.interp_path()): {env.is_pyz(env.interp_path(), debug=debug, suppress_debug=True)}")
+    print(f"is_pipx(env.interp_path()): {env.is_pipx(env.interp_path(), debug=debug, suppress_debug=True)}")
+    print(f"is_python_script(env.interp_path()): {env.is_python_script(env.interp_path(), debug=debug, suppress_debug=True)}")
     print("\nCurrent Environment Check")
     print("# // Based on sys.argv[0]")
     print("-----------------------------")
@@ -85,15 +87,15 @@ def report(path=None, debug=False):
             # so that only unique information gets printed for each check, 
             # even when more than one use the same functions which include debugging logs.
             #print(f"check_executable_path(script_path, debug=True)")
-            check_executable_path(script_path, debug=debug)
+            env.check_executable_path(script_path, debug=debug)
             #print(f"read_magic_bites(script_path, debug=True)")
-            read_magic_bytes(script_path, debug=debug)
-        print(f"is_elf(): {is_elf(script_path, debug=debug, suppress_debug=True)}")
-        print(f"is_windows_portable_executable(): {is_windows_portable_executable(script_path, debug=debug, suppress_debug=True)}")
-        print(f"is_macos_executable(): {is_macos_executable(script_path, debug=debug, suppress_debug=True)}")
-        print(f"is_pyz(): {is_pyz(script_path, debug=debug, suppress_debug=True)}")
-        print(f"is_pipx(): {is_pipx(script_path, debug=debug, suppress_debug=True)}")
-        print(f"is_python_script(): {is_python_script(script_path, debug=debug, suppress_debug=True)}")
+            env.read_magic_bytes(script_path, debug=debug)
+        print(f"is_elf(): {env.is_elf(script_path, debug=debug, suppress_debug=True)}")
+        print(f"is_windows_portable_executable(): {env.is_windows_portable_executable(script_path, debug=debug, suppress_debug=True)}")
+        print(f"is_macos_executable(): {env.is_macos_executable(script_path, debug=debug, suppress_debug=True)}")
+        print(f"is_pyz(): {env.is_pyz(script_path, debug=debug, suppress_debug=True)}")
+        print(f"is_pipx(): {env.is_pipx(script_path, debug=debug, suppress_debug=True)}")
+        print(f"is_python_script(): {env.is_python_script(script_path, debug=debug, suppress_debug=True)}")
     else:
         print("Skipping: ") 
         print("    is_elf(), ")
@@ -108,7 +110,7 @@ def report(path=None, debug=False):
     print("=== PyHabitat Report Complete ===")
     print("=================================")
     print("")
-    interactive = in_repl() or sys.flags.interactive
+    interactive = env.in_repl() or sys.flags.interactive
     if not interactive:
         # Keep window open.
         try: